IT LIVES
Corrected a minor error re: initializing variable i

Code runs more or less correctly. It does not correctly display the
gallows but it does let you know if you entered word correctly. And if
you win (<10 tries to guess), it displays the correct gallows.
Things missing: error checking of any sort (input is assumed to be
correct. there is no sanitization of player2 input so it must be Exactly
As Player 1 Wrote)

Player 1 - Enter a word (max 10 characters): cat
--------------------------------------------
| #  #   #   #   #  #### #   #   #   #   # |
| #  #  # #  ##  # #     ## ##  # #  ##  # |
| #### ##### # # # #  ## # # # ##### # # # |
| #  # #   # #  ## #   # #   # #   # #  ## |
| #  # #   # #   #  ###  #   # #   # #   # |
--------------------------------------------

Welcome to the game Hangman!

Player 1 enters a word and Player 2 guesses it.
Player 2 enters a single lower or upper-case character.
If 10 incorrect guesses have been made, Player 1 wins.

Player 2 - This is the word to guess: ...
c

That letter was correct.

The word including the letters you guessed: c..

Amount of wrong letters: 0

____________

1.     Enter the letter(s) you want to guess:
That letter was incorrect.

The word including the letters you guessed: c..

Amount of wrong letters: 1

|
|
|
|
|
__|_________

2.     Enter the letter(s) you want to guess: t

That letter was correct.

The word including the letters you guessed: c.t

Amount of wrong letters: 1

|
|
|
|
|
__|_________

3.     Enter the letter(s) you want to guess:
That letter was incorrect.

The word including the letters you guessed: c.t

Amount of wrong letters: 2

_______
|
|
|
|
|
__|_________

4.     Enter the letter(s) you want to guess: u

That letter was incorrect.

The word including the letters you guessed: c.t

Amount of wrong letters: 3

_______
|/
|
|
|
|
__|_________

5.     Enter the letter(s) you want to guess:
That letter was incorrect.

The word including the letters you guessed: c.t

Amount of wrong letters: 4

_______
|/   |
|    O
|
|
|
__|_________

6.     Enter the letter(s) you want to guess: a

That letter was correct.

The word including the letters you guessed: cat

Amount of wrong letters: 4

_______
|/   |
|    O
|
|
|
__|_________